Refactor ReceivePacket proof fields
