+++ crates/relayer-cosmos/src/impls/chain.rs
@@ -57,8 +57,7 @@ use crate::methods::packet::{
     build_ack_packet_message, build_ack_packet_payload, build_receive_packet_message,
     build_receive_packet_payload, build_timeout_unordered_packet_message,
     build_timeout_unordered_packet_payload, query_is_packet_received,
-    query_write_acknowledgement_event, CosmosAckPacketPayload, CosmosReceivePacketPayload,
-    CosmosTimeoutUnorderedPacketPayload,
+    query_write_acknowledgement_event,
 };
 use crate::methods::update_client::{build_update_client_message, build_update_client_payload};
 use crate::traits::message::CosmosMessage;
@@ -76,6 +75,9 @@ use crate::types::error::{BaseError, Error};
 use crate::types::payloads::connection::{
     CosmosConnectionOpenAckPayload, CosmosConnectionOpenConfirmPayload,
     CosmosConnectionOpenInitPayload, CosmosConnectionOpenTryPayload,
+};
+use crate::types::payloads::packet::{
+    CosmosAckPacketPayload, CosmosReceivePacketPayload, CosmosTimeoutUnorderedPacketPayload,
 };
 use crate::types::telemetry::CosmosTelemetry;
